scripts/requirements.txt
# Artha runtime dependencies
# ─────────────────────────────────────────────────────────────────────────────
# Quick install (all connectors + skills):
#   pip install -r scripts/requirements.txt
#
# Or install only what you need via pyproject.toml extras:
#   pip install -e ".[google]"        # Gmail + Google Calendar only
#   pip install -e ".[microsoft]"     # Outlook + Microsoft To Do only
#   pip install -e ".[apple]"         # iCloud Mail + Calendar only
#   pip install -e ".[skills]"        # Background skills (NHTSA, weather…)
#   pip install -e ".[all]"           # All connectors and skills
#   pip install -e ".[all,dev]"       # Everything + test tools
#
# All packages are official or widely-trusted; no third-party MCP servers.

# ── Core (always required) ────────────────────────────────────────────────────
PyYAML==6.0.2
keyring==25.6.0
jsonschema==4.23.0
pydantic>=2.0.0

# ── Google connector ─────────────────────────────────────────────────────────
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# ── Microsoft connector ───────────────────────────────────────────────────────
msal==1.32.0
requests==2.32.3

# ── Apple / iCloud connector ──────────────────────────────────────────────────
caldav==1.4.0
vobject==0.9.8

# ── Skills (background data pulls) ────────────────────────────────────────────
beautifulsoup4==4.12.3

# ── Dev / test tools (included here so CI works with requirements.txt) ────────
# For production-only installs, use: pip install -e ".[all]" (skips these)
anyio[trio]>=4.0.0   # required for @pytest.mark.anyio in async tests
# pytest>=7.4.0       ← install separately or via: pip install -e ".[dev]"
# pytest-mock>=3.11.0
# pytest-snapshot>=0.9.0
# datadiff>=2.2.0
